refactor(idea_service): extract pure text helpers (#163)
normalize_tags, validate_raw_tags, slugify (and the _TAG_SLUG_PATTERN
they share) extracted to idea_text_helpers.py — pure functions with
no idea-state dependencies. Re-imported into idea_service for backward
compat (external callers use idea_service.normalize_tags etc.).

  before: idea_service.py = 2524 lines
  after:  idea_service.py = 2481 lines (-43)
          idea_text_helpers.py = 67 lines (new)

321 tests pass. First of several extractions to address modularity
drift in this file. Each subsequent breath continues the work.
